core: use switch.tsx
