+++ static/app/components/core/switch/switch.chonk.tsx
