+++ static/app/components/core/switch/switch.spec.tsx
